requirements.txt
flask-script==2.0.5
flask-sqlalchemy==2.1
aiohttp==0.18.4
